add a lot more error handling

+++ bin/1.18/behat-prepare-simplesaml1.18.0.sh
@@ -72,9 +72,38 @@ rsync -av \
 	--exclude='.git' \
 	./* "$PREPARE_DIR"/wp-content/plugins/wp-saml-auth
 
+# Check that the CIRCLE_WORKING _DIRECTORY exists
+if [ ! -d "$CIRCLE_WORKING_DIRECTORY" ]; then
+	echo "CIRCLE_WORKING_DIRECTORY ($CIRCLE_WORKING_DIRECTORY) does not exist"
+	exit 1
+fi
+
+# Check that "$BASH_DIR"/1-adminnotice.feature exists.
+if [ ! -f "$BASH_DIR"/1-adminnotice.feature ]; then
+	echo "\"$BASH_DIR/1-adminnotice.feature\" does not exist"
+	exit 1
+fi
+
+# Check that $CIRCLE_WORKING_DIRECTORY/tests exists
+if [ ! -d "$CIRCLE_WORKING_DIRECTORY/tests" ]; then
+	echo "$CIRCLE_WORKING_DIRECTORY/tests does not exist"
+	exit 1
+fi
+
+# Check that $CIRCLE_WORKING_DIRECTORY/tests contains a behat directory
+if [ ! -d "$CIRCLE_WORKING_DIRECTORY/tests/behat" ]; then
+	echo "$CIRCLE_WORKING_DIRECTORY/tests/behat does not exist"
+	exit 1
+fi
+
+# Check that $CIRCLE_WORKING_DIRECTORY/tests/behat contains 0-login.feature
+if [ ! -f "$CIRCLE_WORKING_DIRECTORY/tests/behat/0-login.feature" ]; then
+	echo "$CIRCLE_WORKING_DIRECTORY/tests/behat/0-login.feature does not exist"
+	exit 1
+fi
+
+# If we got through all that stuff, we should be good to copy the file now.
 echo "Copying 1-adminnotice.feature to local Behat tests directory (${CIRCLE_WORKING_DIRECTORY}/tests/behat/)"
-# $BASH_DIR is the absolute path to bin/1.18/
-# PROJECT_ROOT is derived as two levels up from BASH_DIR
 cp "$BASH_DIR"/1-adminnotice.feature "$CIRCLE_WORKING_DIRECTORY"/tests/behat/
 
 ###
